refactor(retweet): merge retweetjs into tweetjs
