--- ApiDesign/twitter/controllers/retweet.js
@@ -1,14 +1,0 @@
-
-const twitter = require('../modules');
-
-function retweet(request, response) {
-    var id = request.params.id;
-    twitter.retweet.retweet(id);
-}
-
-
-module.exports = {
-
-retweet : retweet
-
-}
